ronnia/requirements.txt
twitchio==2.1.5
irc>=19.0.0,<21.0.0
aiosqlite==0.17.0
azure-servicebus==7.6.0